Bonus implemented

+++ app/controllers/authors/author.js
@@ -3,10 +3,13 @@ import Controller from '@ember/controller';
 export default Controller.extend({
   actions: {
     deletePost(post) {
-      post.destroyRecord().then(() => {
-        this.get('model').posts.removeObject(post);
-        this.get('model').save();
-      });
+      let deleteConfirmation = confirm('Are you sure?');
+      if (deleteConfirmation) {
+        post.destroyRecord().then(() => {
+          this.get('model').posts.removeObject(post);
+          this.get('model').save();
+        });
+      }
     }
   }
 });
